requirements.txt
pandas
numpy
Pillow
matplotlib